Create README.md
GAULES COIN THE MEMECOIN REVOLUTION HAS BEGUN!
THE FRENCH ROOSTER! THE MEMECOIN REVOLUTION HAS BEGUN!
Are you ready to be part of the next big cryptocurrency trend?
GAULES COIN (GLS) is more than just a memecoin. It is an ambitious project that is poised to take the market by storm, with a complete ecosystem and a clear growth plan. Our goal is simple: Top 10 and soon Top 1 on exchanges worldwide!